test_app.py
import unittest
import os
import tempfile
from flask_testing import TestCase
from app import PDFSummarizer

class FlaskAppTestCase(TestCase):
    TESTING = True

    def create_app(self):
        # Create a temporary database for testing
        db_fd, db_path = tempfile.mkstemp()
        self.db_fd = db_fd
        self.db_path = db_path

        self.flask_app = PDFSummarizer(database=db_path)
        self.flask_app.app.config['TESTING'] = True
        return self.flask_app.app

    def setUp(self):
        # Set up the test client and initialize the database
        self.app = self.create_app().test_client()
        with self.flask_app.app.app_context():
            self.flask_app.init_db()

    def tearDown(self):
        # Close and remove the temporary database
        os.close(self.db_fd)
        os.unlink(self.db_path)

    def upload_file(self, filename):
        # Helper function to simulate file upload
        try:
            with open(filename, 'rb') as f:
                data = {'file': (f, os.path.basename(filename))}
                return self.app.post('/upload', data=data, content_type='multipart/form-data')
        except FileNotFoundError:
            return None

    # Complete Test Case for PDF Upload
    def test_pdf_upload(self):
        response = self.upload_file('test_data/sample.pdf')
        self.assertEqual(response.status_code, 200)

        response = self.upload_file('test_data/sample.txt')
        if response:
            self.assertNotEqual(response.status_code, 200)

    # Placeholder for Summary Generation Test
    def test_summary_generation(self):
        # TODO: Implement a test to check if the summary is correctly generated.
        # You should upload a PDF and assert that the summary is included in the response.
        pass

    # Placeholder for Exception Handling Test
    def test_exception_handling(self):
        # TODO: Implement a test to check how the application handles exceptions.
        # Consider scenarios such as uploading an invalid file or causing an internal error.
        pass

    # Placeholder for Database Cache Test
    def test_database_cache(self):
        # TODO: Implement a test to verify that the database caching is working.
        # You should upload the same PDF twice and check if the second upload fetches the summary from the cache.
        pass

if __name__ == '__main__':
    unittest.main()

